Dharanish/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==24.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
